+++ srcs/utils/math_utils.c
@@ -6,7 +6,7 @@
 /*   By: mmaquine <mmaquine@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/30 17:38:21 by mmaquine          #+#    #+#             */
-/*   Updated: 2026/03/30 17:51:41 by mmaquine         ###   ########.fr       */
+/*   Updated: 2026/04/07 11:24:50 by gabrgarc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 /*
 Converts degree to radians
 */
-double to_radians(double degree)
+REAL to_radians(REAL degree)
 {
 	return (degree * (M_PI / 180.0));
 }
